requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
flask==2.3.3
matplotlib==3.7.2
pillow==10.0.0
scikit-learn==1.3.0
werkzeug==2.3.7
python-dotenv==1.0.0 